Added S4 stuff along with skeleton documentation.
Fixed usage section in man page for merge.data.frame.

The S4 addition doesn't break anything, and once we figure out best way to test it we can advertise it better by adding it to vignettes and cross-linking the S4 documentation from within the See Also section of ?data.table.

I just wanted to get it into SVN for now.