Fixed AllS4.R and tests for S4 stuff.
The setClass("S4Composition", representation(info="data.table")) was part of a test case to see if we can make an S4 class that has a data.table slot, then define methods that work on that data.table.

This was moved out of AllS4.R and into tests/tests.R. Things seem kosher now.