Ensure .so install name & filename match on OSX (fixes #1144)
OS X convention requires that the basename part of a shared library's
install name (i.e., install name minus directory components) match the
actual filename of said shared library. Otherwise, the dynamic linker
(and related tools) could get confused when trying to find, process,
and/or load the dynamic library.

On OS X, the "mv" command at the end of the build process changes the
shared library filename to "datatable.so" while leaving the install name
as "data.table.so", so we need to run install_name_tool(1) to fix the
install name to match the filename. Currently not a big issue for most
users (since not much else links to this shared library), but not doing
this causes issues with build systems (e.g., conda) that end up
modifying the .so's RPATH for some reason.