Need note about set2key name change
I was confused about `Error: could not find function "setindex"` when I read this vignette first. After some search it seemed to be an new name only available in newer version of data.table.

Since many reader are using the CRAN version 1.9.6, it's better to give a note about the name change.

My edit is just a placeholder, not necessarily in the best format.