Test numbering tidied
Currently, we have to skip test numbers ending zero because, for example, if 1419.10 fails, it prints as 1419.1 the same as 1419.1 and we can't distinguish.  Haven't thought of a solution yet.  For now, I just think of it as a bit of culture/personality in the code,  like missing unlucky number 13.