Fix file size display bug in fread
The bug only manifests for files of size > 1TB. Which is why it was not spotted earlier, and why it's impossible to write tests for it.

The problem was that the expression `(1 << 40)` was interpreted as `int(1) << 40`, causing an overflow, and as a result incorrect file size was displayed in the output. On the other hand `(1LL << 40)` is `long long int(1) << 40`, which no longer overflows.