Add macro HAVE_LIBZ in fwrite
HAVE_LIBZ must be defined if zlib (lib and headers) is present on the
system. If not,  HAVE_LIBZ is undefined and data.table is compiled
without zlib support.

Normally  HAVE_LIBZ (or an equivalent) will be defined (or not) by a configure
script (see ax_check_zlib).

See issue #3872.