Improved multi-column handling in as.matrix
Elimanted costly check of each column for dimensions, replacing it with a single (very fast) call to setDT to check for columns with multiple columns (e.g. a column that contains a matrix).

In the rare edge case where this is detected (see test 2074.07) we now use as.data.table to unpack these columns.

This elimanted the need for the later check for matrix and data.table columns and subsequent use of the old unlist rather than the new C method.

Test 2074.07 required minor modification in the expected column names of the output matrix.