homogenous non-atomic types converted correctly
When a data.table contains columns all of the same class, the matrix will also be given that class.

This means that classes that extend atomic types will work correctly. For example, integer64 in the bit64 package, which is a class attached to a numeric type, will now work correctly, where before a numeric matrix was returned without the appropriate class thus the values were at face value nonsensical. Note this was a problem even for the old unlist method.