Split out handling of class and types
integer64 is now treated as a special case when checking column classes, and any future classes should also be handled in this manner.

type checking is then used as a fallback method; columns are coerced to a common type if necessary (i.e. where encountering classes that are not from base R or one of the specially handled cases).