Improved memory when list + factor or date
Moved code for converting factor columns and Date/POSIX columns to character to happen prior to checking for list/expression/recursive columns. This improves memory usage where these columns co-occur with recursive columns, as it reduces the information stored in each list element cell of the matrix (e.g. before all factor levels would be stored as an attribute in each cell of the matrix corresponding to the factor column).