Use consistent types with fwriteMainArgs.nrow.
On armv7hl, test 1737.5 fails due to garbage in the column length
mismatch error message. This is because the message tries to format
`args.nrow` (an `int64_t`) using `%d` (i.e., `int`). Strangely, this
does not fail on any other architectures, but this is likely a fluke.

In all the other formatting calls, remove the unnecessary `(int64_t)`
type cast, since `fwriteMainArgs.nargs` already is one.