Tests for .ITime: round, trunc, floor, and ceiling
Tests should be extended for correctness. 
One known mistake is in round() when second == 30 since this will make division == x.5, which means that it will round up or down. There is no logic for when it does what. But this is a general problem with the round() function, don't know how to address this.