add `datatable.fwrite.sep` option
With many kind of data, particularly with data heavy in text
fields, using ";" as a default separator can cause issues
when imported with naive parsers. This commit add a new global
option to set the default fwrite separator. Even with this option set,
the separator can still be changed by using the `sep` argument. When not
set, the default separator is still ",".