Update datatable-intro.Rmd
A powerful feature of data.table is the ability to store lists as elements (a facility not available in data.frame).  Here is a simple illustration of storing a fitted model as data.table element.

When statistical models are fit to different groups within a data.table, these models can be stored in a result data.table.  Now this resulting data.table can be used to access fitted models.