fread(): avoid casting int* to Rboolean* (#6782)
* fread(): avoid casting int* to Rboolean*

This used to work because the compilers were very likely to use int as
the underlying type of the enum. R-devel r87656 changes the definition
of Rboolean from an enum to bool, sothe sizes and the semantics of the
types are very different now. Since LOGICAL() returns int*, and
NA_INTEGER is an int, keep using that to access the vector contents.

* safe change in %notin%

* safe change sizeof()

* Local logical scalars LOGICAL(x)[0] --> int, not Rboolean

* One more sizeof treatment

---------

Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>
Co-authored-by: Michael Chirico <chiricom@google.com>