Avoid casts that don't change the pointer type (#6785)
* Don't cast LOGICAL() to int*

If the returned pointer type ever changes, the absence of the cast will
give us a compiler warning due to the value being implicitly cast to an
incompatible pointer type. An explicit cast, on the other hand, tells
the compiler to silence the warning and do as written.

* Drop more no-op casts

* Introduce the Coccinelle linter

For mistakes that are easy to fix automatically, Coccinelle patches may
be written. If a patch finds a mistake, the linter will ask the human to
fix either the code or the patch.