Use notOutput= instead of capture.output() tests
Where capture.output() was previously used to test the absence of
messages, use notOutput= instead so that the test will be skipped.

While the test succeeded anyway due to the captured output containing
neither the original untranslated message (being matched) nor its
translation (that could be printed if a regression happened), skipping
the test is more fair.