Avoid a memory leak in fwrite(compress="gzip")
Previously, the zlib object zstrm was initialized whenever compression
was enabled, at least in order to calculate the buffer size for the
compressed data stream, but was only used for compression and
subsequently deallocated only when there was a header with column names.

Make sure to deflateEnd(&strm) even if there was no header.