`fwrite()`: when `col.names=FALSE`, produce a `gzip` header, avoid a leak (#6853)
* fwrite(): emit gzip header even if !col.names

Previously, fwrite() emitted a raw DEFLATE stream if col.names was FALSE
(and thus headerLen was zero). Decompressing such files is possible, but
not trivial.

* Avoid a memory leak in fwrite(compress="gzip")

Previously, the zlib object zstrm was initialized whenever compression
was enabled, at least in order to calculate the buffer size for the
compressed data stream, but was only used for compression and
subsequently deallocated only when there was a header with column names.

Make sure to deflateEnd(&strm) even if there was no header.

* NEWS entry

* comment the source of this #endif for clarity

* Clean up the test file

Co-Authored-By: Michael Chirico <chiricom@google.com>

* Report write() errno instead of just plain -1

Co-Authored-By: Michael Chirico <chiricom@google.com>

---------

Co-authored-by: Michael Chirico <chiricom@google.com>