`fwrite`: pre-encode strings and factor levels (#6889)
* fwrite: pre-encode strings and factor levels

Previously, fwrite() deferred encoding of the strings to
fwriteR.c:getString,getCategString called from OpenMP threads. Calling
translateChar[UTF8] to encode a string results in memory allocation
unless it is already in the desired encoding, which is unsafe to perform
on a non-main thread.

Fixes: #6883

* tests: fix iconv() source encoding

Since iconv() ignores the encoding bits, we must provide the correct
from=... argument. The default from="" would only work with a UTF-8
locale. Instead, assume that "\uXX" strings are UTF-8-encoded.

* Fix test

Use test data that is more likely to reproduce the crash. Fix the number
of threads, too.

Co-Authored-By: Benjamin Schwendinger <benjaminschwe@gmail.com>

* NEWS entry

---------

Co-authored-by: Benjamin Schwendinger <benjaminschwe@gmail.com>