fread: better error for files that are too large
On non-Windows, print the errno when file exists but we couldn't open
it. That will happen for files larger than the address space when off_t
is 32-bit (default on Linux+glibc).

In case an operating system does let us open a large file, check the
size before casting it to size_t. (This happens at least on Windows.)