Fix warnings and truncations around `filesize_to_str()` (#6987)
* filesize_to_str(): accept uint64_t, not size_t

On 32-bit platforms, we would otherwise perform a 40-bit shift of a
32-bit size_t, which is undefined behaviour. Moreover, files can be
larger than the address space of the process.

* fread: better error for files that are too large

On non-Windows, print the errno when file exists but we couldn't open
it. That will happen for files larger than the address space when off_t
is 32-bit (default on Linux+glibc).

In case an operating system does let us open a large file, check the
size before casting it to size_t. (This happens at least on Windows.)

* filesize_to_str(): print binary SI prefixes

* Update src/fread.c

Co-authored-by: Michael Chirico <chiricom@google.com>

* Print both file size and SIZE_MAX upon failure

Co-authored-by: Michael Chirico <chiricom@google.com>

* NEWS entries

* Fix double blunder

- Make sure to use the right platform-specific variable in both cases
- Can't use filesize_to_str() more than once due to its static buffer

* tweak NEWS

---------

Co-authored-by: Michael Chirico <chiricom@google.com>
Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>