mirror.packages: download newer versions
Only skip downloading a package if it already has the right version and
MD5sum (in case of binary rebuilds). When downloading a new version of
the package, remove the old one.

This should help ensure that the source packages mirrored from CRAN are
as new as binaries (or newer), which install.packages(type = "both")
assumes when resolving dependencies.