Fixes for `integer64` issues in `between` (#7193)
* tests: failing integer64 readable as small integer

* between: check for integer64 before coercion

In particular, avoid calling fitsInInt32() on REALSXP vectors that could
possibly be of class 'integer64' (and therefore shouldn't be read as
doubles).

* fitInInt*: disallow integer64

Since the functions themselves read the vector as double* and are used
as a test whether doubles from the vector would fit in an integer, avoid
reading integer64 vectors altogether.

* tests: failing blind double -> int64 conversion

* between: convert bounds to int64 only if they fit

Avoid auto-converting 'lower' or 'upper' to integer64 if they aren't
actually representable in int64_t.

* Apply suggestions from code review

Unify the translatable strings.

Co-authored-by: Michael Chirico <chiricom@google.com>

* Fix tests for error messages

* Make the error messages more symmetric.

Co-authored-by: Michael Chirico <chiricom@google.com>

---------

Co-authored-by: Michael Chirico <chiricom@google.com>