Migrate most uses of `SETLENGTH` to the resizable API (#7451)
* Use the experimental resizable vectors API

Thanks to Luke Tierney for introducing the API and helping with the migration.

* Backport the resizable API

Make sure to set the GROWABLE_BIT on the resizable vectors to avoid
problems when they are duplicated or garbage-collected.

* test 2291.1: misleading TRUELENGTH now impossible

Now that data.table objects have the GROWABLE_BIT set, R will reset TRUELENGTH
when duplicating them, causing our code to take a different branch.

* Drop the finalizer

Now that (1) we depend on R >= 3.4 and (2) data.table objects have the
GROWABLE_BIT set, there is no need to adjust allocated memory counts by
hand.

* frollapply(adaptive=TRUE): resizable temporaries

Since adaptive application of rolling functions requires us to resize
the argument to match the window size, make sure to allocate it as such.

* Drop remaining uses of TRUELENGTH from assign.c

- Don't SET_TRUELENGTH by hand. All of our resizable vectors now have
  the GROWABLE_BIT set, so when they are duplicated, TRUELENGTH is reset
  to 0.
- Use a combination of R_isResizable and R_maxLength to replace other
  uses of TRUELENGTH.

* Drop test for TRUELENGTH from init.c

* Better backport of R_isResizable()

* Placate rchk

* copyAsGrowable: don't crash on 0-len argument

* Move the resizable allocation functions to utils.c

* duplicateAsResizable: refuse ALTREP objects

* maxLength: return xlength if non-resizable

That's what the function does in R-devel.

* adjust to previous code

* Safety checks in R_resizeVector() backport

* fix comment

* Mark internal errors as # nocov

---------

Co-authored-by: Jan Gorecki <j.gorecki@wit.edu.pl>
Co-authored-by: Benjamin Schwendinger <52290390+ben-schwen@users.noreply.github.com>