Interruption problems in `frollapply` (#7428)
* frollapply: let the interrupt continue

Instead of calling invokeRestart("abort") in the interrupt handler,
return from it. This continues the dispatch of the interrupt and lets an
outer handler catch it:

tryCatch(
  frollapply(1:1e6, 1, \(.) { Sys.sleep(ret <- sum(.)); ret}),
  interrupt = \(e) 'interrupted'
)
^C[1] "interrupted"

With invokeRestart("abort"), the interrupt cannot be handled further.

* frollapply: wait for terminated parallel processes

While handling an interrupt, ask mccollect() to wait for the child
process to exit (with a warning) in order to avoid producing zombies.
Otherwise a process that is too slow to react to SIGTERM will remain a
zombie until the parent process exits.

* Use waitid() to choose processes to terminate

Since the PIDs of our worker processes could have been reused, first
test them using waitid(NOWAIT) to make sure they are still ours.

* Interruption handling is not covered