More test fixes for when translations are present (#7478)
* Switch more tests to test(output=...)

capture.output() tests are fine in many places, but these ones tested
for messages that have since been translated and must be skipped in
foreign more.

* Use notOutput= instead of capture.output() tests

Where capture.output() was previously used to test the absence of
messages, use notOutput= instead so that the test will be skipped.

While the test succeeded anyway due to the captured output containing
neither the original untranslated message (being matched) nor its
translation (that could be printed if a regression happened), skipping
the test is more fair.

* test.data.table(): count skipped message tests

When working in foreign mode, count the occurrences of tests for message
content (which are essentially skipped) and report the number if it's
non-zero.

* test 1590.14: don't assume 1252 if not UTF-8

* Elaborate on what 1252 is

* Better test for not printing NULL

* Restore test 1775.1 for print() options

Instead, test substitute(x) for %iscall% "print" to decide whether to
print the x before testing the output.